+++ src/lib/index.ts
@@ -30,7 +30,7 @@ import { getPackageInfo } from "../utils/getPackageInfo";
 import { sendPayment, registerPaymentUser } from "../utils/sendPayment";
 import { handleUrlQuery } from "../utils/handleUrlQuery";
 import { prepareProviders, renderGooglePayButton } from "../utils/loadProviderSdks";
-import { createAgreementModal, createPaymentSuccessForm } from "./create";
+import { createPaymentSuccessForm } from "./create";
 import { CheckoutAPI } from "../utils/api";
 import { Logger } from './logger';
 import { getFormValues, loadSelectbox } from "./common";
@@ -52,7 +52,8 @@ async function ZotloCheckout(params: IZotloCheckoutParams): Promise<IZotloChecko
       style: params.style,
       customParameters: params.customParameters,
       useCookie: !!params?.useCookie,
-      showSavedCards: params?.showSavedCards
+      showSavedCards: params?.showSavedCards,
+      quantitySetting: params?.quantitySetting,
     });
     await refreshProviderConfigs();
   }
@@ -60,7 +61,6 @@ async function ZotloCheckout(params: IZotloCheckoutParams): Promise<IZotloChecko
   const maskItems: Record<string, ReturnType<typeof maskInput>> = {};
   const validations: Record<string, ReturnType<typeof validateInput>> = {};
   const selectboxList: Record<string, ReturnType<typeof loadSelectbox>> = {};
-  let destroyAgreementLinks = null as (() => void) | null;
   let destroySavedCardsEvents = null as (() => void) | null;
   let destroyBillingFormEvents = null as (() => void) | null;
 
@@ -133,66 +133,6 @@ async function ZotloCheckout(params: IZotloCheckoutParams): Promise<IZotloChecko
 
   function hasAnyConfig() {
     return Object.keys(config.settings).length > 0;
-  }
-
-  function handleAgreementLinks() {
-    const container = ZOTLO_GLOBAL.container?.querySelector('form');
-    const buttons = container?.querySelectorAll('[data-agreement]') as NodeListOf<HTMLButtonElement>;    
-
-    function closeAgreement() {
-      container?.querySelector('[data-modal="agreement"]')?.remove();
-    }
-
-    function handleClick(this: HTMLElement) {
-      try {
-        const target = this as HTMLElement;
-        const agreement = target.getAttribute('data-agreement') as any;
-        const modalHTML = createAgreementModal({ key: agreement, config })
-        const parser = new DOMParser();
-        let modalDOM = parser.parseFromString(modalHTML, 'text/html')?.body.firstChild as HTMLElement;
-
-        // Add modal close action
-        modalDOM?.querySelector('[data-modal-close]')?.addEventListener('click', handleClose);
-
-        container?.insertBefore(modalDOM, container.firstChild as HTMLElement);
-        modalDOM = container?.querySelector(`[data-modal="agreement"]`) as HTMLElement;
-
-        setTimeout(() => {
-          modalDOM?.classList.remove('zotlo-checkout__modal-enter-from');
-          modalDOM?.classList.remove('zotlo-checkout__modal-enter-active');
-        }, 0)
-
-        function handleClose(this: HTMLElement) {
-          const closeBtn = this as HTMLElement;
-          modalDOM?.classList.add('zotlo-checkout__modal-enter-from');
-          modalDOM?.classList.add('zotlo-checkout__modal-enter-active');
-          closeBtn.removeEventListener('click', handleClose);
-          
-          setTimeout(() => closeAgreement(), 150);
-        }
-      } catch (e) {
-        Logger.client?.captureException(e);
-      }
-    }
-
-    if (buttons?.length > 0) {
-      for (const button of buttons) {
-        button.addEventListener('click', handleClick);
-      }
-    }
-
-    function destroy() {
-      closeAgreement();
-      if (!buttons || buttons?.length === 0) return;
-
-      for (const button of buttons) {
-        button.removeEventListener('click', handleClick);
-      }
-    }
-
-    return {
-      destroy
-    }
   }
 
   function handleTabView() {
@@ -479,6 +419,18 @@ async function ZotloCheckout(params: IZotloCheckoutParams): Promise<IZotloChecko
                   result
                 });
               }
+
+              // Set state for subscriber ID and zip code
+              if ([
+                FORM_ITEMS.SUBSCRIBER_ID_EMAIL.input.name,
+                FORM_ITEMS.SUBSCRIBER_ID_PHONE.input.name
+              ].includes(item.name)) {
+                ZOTLO_GLOBAL.data.subscriberId = item.value;
+              }
+
+              if (item.name === FORM_ITEMS.ZIP_CODE.input.name) {
+                ZOTLO_GLOBAL.data.zipCode = item.value;
+              }
             }
           });
         }
@@ -488,9 +440,7 @@ async function ZotloCheckout(params: IZotloCheckoutParams): Promise<IZotloChecko
     applyMaskAndValidation();
 
     if (import.meta.env.VITE_SDK_API_URL) {
-      const { destroy } = handleAgreementLinks();
       const { destroy: destroyFn } = handleSavedCardsEvents({ config });
-      destroyAgreementLinks = destroy;
       destroySavedCardsEvents = destroyFn;
       renderGooglePayButton(config);
     }
@@ -541,7 +491,6 @@ async function ZotloCheckout(params: IZotloCheckoutParams): Promise<IZotloChecko
     }
 
     validatorInstance?.clearRules();
-    destroyAgreementLinks?.();
     destroySavedCardsEvents?.();
     destroyBillingFormEvents?.();
   }
@@ -572,7 +521,7 @@ async function ZotloCheckout(params: IZotloCheckoutParams): Promise<IZotloChecko
     destroyFormInputs();
     const container = ZOTLO_GLOBAL.container;
     if (container) container.innerHTML = '';
-    ZOTLO_GLOBAL.containerId = '';
+    ZOTLO_GLOBAL.reset();
   }
 
   function mount(id: string) {