+++ src/lib/theme/mobileapp/index.ts
@@ -2,8 +2,8 @@ import mainHTML from './html/main.html?raw';
 import { generateAttributes, getCDNUrl, useI18n } from '../../../utils'
 import { template } from "../../../utils/template";
 import { PaymentProvider, type FormConfig, type FormSetting } from '../../types';
-import { createButton, createCreditCardForm, createPaymentHeader } from '../../create'
-import { getPackageName } from '../../../utils/getPackageInfo';
+import { createButton, createCreditCardForm, createFooter, createPaymentHeader } from '../../create'
+import { getPackageName, getQuantityInfo } from '../../../utils/getPackageInfo';
 import { prepareProvider } from './utils';
 
 export function generateThemeMobileApp(params: {
@@ -15,8 +15,8 @@ export function generateThemeMobileApp(params: {
     PRICE_INFO: string;
     FOOTER_DESC: string;
     DISCLAIMER: string;
-    ZOTLO_LEGALS_DESC: string;
-    ZOTLO_LEGALS_LINKS: string
+    ZOTLO_LEGALS_TEXT: string;
+    PAYMENT_AGGREGATOR: string;
   };
 }) {
   const { config, dir, themePreference, paymentMethods, footerInfo } = params;
@@ -103,6 +103,7 @@ export function generateThemeMobileApp(params: {
     : '';
 
   const paymentHeader = createPaymentHeader({ config });
+  const footer = createFooter(footerInfo) || '';
 
   return template(mainHTML, {
     DIR: dir,
@@ -122,13 +123,10 @@ export function generateThemeMobileApp(params: {
     ADDITIONAL_TEXT: additionalText,
     ADDITIONAL_PRICE: additionalPrice,
     TOTAL_PRICE: totalPrice,
+    QUANTITY_INFO: getQuantityInfo(config),
     PRIMARY_PROVIDER: primaryProvider,
     TAB_BUTTONS: tabButtons,
     PROVIDERS: providerButtons,
-    PRICE_INFO: footerInfo.PRICE_INFO,
-    FOOTER_DESC: footerInfo.FOOTER_DESC,
-    DISCLAIMER: footerInfo.DISCLAIMER,
-    ZOTLO_LEGALS_DESC: footerInfo.ZOTLO_LEGALS_DESC,
-    ZOTLO_LEGALS_LINKS: footerInfo.ZOTLO_LEGALS_LINKS,
+    FOOTER: footer
   })
 }